Implemented support for the vertical panel position.
It's now possible to specify the vertical panel position (Top, Center or Bottom).
We default to 'Bottom' if there isn't any position specified in the *secondary* config.
We *do not* force the panel position if it is Bottom, so if you use a custom tint2 configuration
you can specify Bottom in the panel config and the panel_position variable will not be touched.